+++ test/io_tests/tunnel_test.py
@@ -1,27 +1,33 @@
 """Test for KNX/IP Tunnelling connections."""
 import asyncio
+from copy import deepcopy
 from unittest.mock import AsyncMock, Mock, call, patch
 
 import pytest
 
 from xknx import XKNX
 from xknx.dpt import DPTArray
-from xknx.io import UDPTunnel
+from xknx.exceptions import CommunicationError
+from xknx.io import TCPTunnel, UDPTunnel
+from xknx.io.const import CONNECTIONSTATE_REQUEST_TIMEOUT, HEARTBEAT_RATE
 from xknx.io.gateway_scanner import GatewayDescriptor
 from xknx.knxip import (
     HPAI,
     CEMIFrame,
+    ConnectionStateRequest,
+    ConnectionStateResponse,
     ConnectRequest,
     ConnectResponse,
     DescriptionRequest,
     DescriptionResponse,
     DisconnectRequest,
     DisconnectResponse,
+    ErrorCode,
     KNXIPFrame,
     TunnellingAck,
     TunnellingRequest,
 )
-from xknx.knxip.knxip_enum import CEMIMessageCode
+from xknx.knxip.knxip_enum import CEMIMessageCode, HostProtocol
 from xknx.telegram import IndividualAddress, Telegram, TelegramDirection
 from xknx.telegram.apci import GroupValueWrite
 
@@ -53,8 +59,8 @@ class TestUDPTunnel:
             # communication_channel_id: 0x02   sequence_counter: 0x21
             bytes.fromhex("0610 0420 0017 04 02 21 00 2900bcd011162916030080 0c 3f"),
             # L_Data.ind T_Connect from 1.0.250 to 1.0.255 (xknx tunnel endpoint)
-            # communication_channel_id: 0x02   sequence_counter: 0x81
-            bytes.fromhex("0610 0420 0014 04 02 81 00 2900b06010fa10ff0080"),
+            # communication_channel_id: 0x02   sequence_counter: 0x21
+            bytes.fromhex("0610 0420 0014 04 02 21 00 2900b06010fa10ff0080"),
         ],
     )
     @patch("xknx.io.UDPTunnel._send_tunnelling_ack")
@@ -64,6 +70,7 @@ class TestUDPTunnel:
         _cemi.from_knx(raw[10:])
         telegram = _cemi.telegram
         telegram.direction = TelegramDirection.INCOMING
+        self.tunnel.expected_sequence_number = 0x21
 
         self.tunnel.transport.data_received_callback(raw, ("192.168.1.2", 3671))
         await asyncio.sleep(0)
@@ -86,6 +93,7 @@ class TestUDPTunnel:
         _cemi.from_knx(raw_req[10:])
         test_telegram = _cemi.telegram
         test_telegram.direction = TelegramDirection.INCOMING
+        self.tunnel.expected_sequence_number = 0x81
 
         response_telegram = Telegram(source_address=self.tunnel._src_address)
 
@@ -111,12 +119,55 @@ class TestUDPTunnel:
         ]
         send_ack_mock.assert_called_once_with(raw_req[7], raw_req[8])
 
+    async def test_repeated_tunnel_request(self, time_travel):
+        """Test Tunnel for receiving repeated TunnellingRequest frames."""
+        self.tunnel.transport.send = Mock()
+        self.tunnel.communication_channel = 1
+        self.tunnel.expected_sequence_number = 10
+
+        test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
+        test_frame = KNXIPFrame.init_from_body(
+            TunnellingRequest(
+                communication_channel_id=1,
+                sequence_counter=10,
+                cemi=CEMIFrame.init_from_telegram(
+                    test_telegram, code=CEMIMessageCode.L_DATA_IND
+                ),
+            )
+        )
+        test_ack = KNXIPFrame.init_from_body(TunnellingAck(sequence_counter=10))
+        test_frame_9 = deepcopy(test_frame)
+        test_frame_9.body.sequence_counter = 9
+
+        # first frame - ACK and processed
+        self.tunnel._request_received(test_frame, None, None)
+        await time_travel(0)
+        assert self.tunnel.transport.send.call_args_list == [call(test_ack, addr=None)]
+        self.tunnel.transport.send.reset_mock()
+        assert self.tunnel.expected_sequence_number == 11
+        assert self.tg_received_mock.call_count == 1
+        # same sequence number as before - ACK, not processed
+        self.tunnel._request_received(test_frame, None, None)
+        await time_travel(0)
+        assert self.tunnel.transport.send.call_args_list == [call(test_ack, addr=None)]
+        self.tunnel.transport.send.reset_mock()
+        assert self.tunnel.expected_sequence_number == 11
+        assert self.tg_received_mock.call_count == 1
+        # wrong sequence number - no ACK, not processed
+        self.tunnel._request_received(test_frame_9, None, None)
+        await time_travel(0)
+        assert self.tunnel.transport.send.call_args_list == []
+        self.tunnel.transport.send.reset_mock()
+        assert self.tunnel.expected_sequence_number == 11
+        assert self.tg_received_mock.call_count == 1
+
     @patch("xknx.io.UDPTunnel._send_tunnelling_ack")
     def test_tunnel_request_received_cemi_too_small(self, send_ack_mock):
         """Test Tunnel sending ACK for unsupported frames."""
         # <UnsupportedCEMIMessage description="CEMI too small. Length: 9; CEMI: 2900b06010fa10ff00" />
         # communication_channel_id: 0x02   sequence_counter: 0x81
         raw = bytes.fromhex("0610 0420 0013 04 02 81 00 2900b06010fa10ff00")
+        self.tunnel.expected_sequence_number = 0x81
 
         self.tunnel.transport.data_received_callback(raw, ("192.168.1.2", 3671))
         self.tg_received_mock.assert_not_called()
@@ -129,15 +180,17 @@ class TestUDPTunnel:
         # <UnsupportedCEMIMessage description="APCI not supported: 0b1111111000 in CEMI: 2900b0d0000100000103f8" />
         # communication_channel_id: 0x02   sequence_counter: 0x4f
         raw = bytes.fromhex("0610 0420 0015 04 02 4f 00 2900b0d0000100000103f8")
+        self.tunnel.expected_sequence_number = 0x4F
 
         self.tunnel.transport.data_received_callback(raw, ("192.168.1.2", 3671))
         self.tg_received_mock.assert_not_called()
         send_ack_mock.assert_called_once_with(0x02, 0x4F)
 
-    async def test_tunnel_wait_for_l2_confirmation(self, time_travel):
+    async def test_tunnel_send_and_wait_for_l2_confirmation(self, time_travel):
         """Test tunnel waits for L_DATA.con before sending another L_DATA.req."""
         self.tunnel.transport.send = Mock()
         self.tunnel.communication_channel = 1
+        self.tunnel.expected_sequence_number = 23
 
         test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
         test_ack = KNXIPFrame.init_from_body(TunnellingAck(sequence_counter=23))
@@ -162,6 +215,69 @@ class TestUDPTunnel:
         # one call for the outgoing request and one for the ACK for the confirmation
         assert self.tunnel.transport.send.call_count == 2
         await task
+
+    async def test_tunnel_send_retry(self, time_travel):
+        """Test tunnel resends the telegram when no ACK was received."""
+        self.tunnel.transport.send = Mock()
+        self.tunnel.communication_channel = 1
+        self.tunnel.sequence_number = 23
+        self.tunnel.expected_sequence_number = 15
+
+        test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
+        request = KNXIPFrame.init_from_body(
+            TunnellingRequest(
+                communication_channel_id=self.tunnel.communication_channel,
+                sequence_counter=self.tunnel.sequence_number,
+                cemi=CEMIFrame.init_from_telegram(
+                    test_telegram,
+                    code=CEMIMessageCode.L_DATA_REQ,
+                    src_addr=self.tunnel._src_address,
+                ),
+            )
+        )
+        test_ack = KNXIPFrame.init_from_body(
+            TunnellingAck(sequence_counter=self.tunnel.sequence_number)
+        )
+        confirmation = KNXIPFrame.init_from_body(
+            TunnellingRequest(
+                communication_channel_id=1,
+                sequence_counter=15,
+                cemi=CEMIFrame.init_from_telegram(
+                    test_telegram, code=CEMIMessageCode.L_DATA_CON
+                ),
+            )
+        )
+        confirmation_ack = KNXIPFrame.init_from_body(TunnellingAck(sequence_counter=15))
+
+        task = asyncio.create_task(self.tunnel.send_telegram(test_telegram))
+        await time_travel(0)
+        assert self.tunnel.transport.send.call_args_list == [call(request, addr=None)]
+        self.tunnel.transport.send.reset_mock()
+        # no ACK received, resend same telegram
+        await time_travel(1)
+        assert self.tunnel.transport.send.call_args_list == [call(request, addr=None)]
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(test_ack, HPAI())
+        await time_travel(0)
+        assert not task.done()
+        self.tunnel.transport.handle_knxipframe(confirmation, HPAI())
+        await time_travel(0)
+        assert task.done()
+        assert self.tunnel.transport.send.call_args_list == [
+            call(confirmation_ack, addr=None)
+        ]
+        self.tunnel.transport.send.reset_mock()
+        await task
+
+        # Test raise after 2 missed ACKs (reconnect if `auto_reconnect` was True)
+        with pytest.raises(CommunicationError):
+            task = asyncio.create_task(self.tunnel.send_telegram(test_telegram))
+            # no ACKs received, for 2x wait time (with advancing the loop in between)
+            await time_travel(1)
+            await time_travel(1)
+            assert self.tunnel.transport.send.call_count == 2
+            self.tunnel.transport.send.reset_mock()
+            await task
 
     @pytest.mark.parametrize(
         "route_back,data_endpoint_addr,local_endpoint",
@@ -270,3 +386,215 @@ class TestUDPTunnel:
         assert isinstance(task.result(), GatewayDescriptor)
         assert self.tunnel.transport.send.call_count == 1
         await task
+
+
+class TestTCPTunnel:
+    """Test class for xknx/io/TCPTunnel objects."""
+
+    def setup_method(self):
+        """Set up test class."""
+        # pylint: disable=attribute-defined-outside-init
+        self.xknx = XKNX()
+        self.tg_received_mock = AsyncMock()
+        self.tunnel = TCPTunnel(
+            self.xknx,
+            gateway_ip="192.168.1.2",
+            gateway_port=3671,
+            telegram_received_callback=self.tg_received_mock,
+            auto_reconnect=False,
+            auto_reconnect_wait=3,
+        )
+
+    async def test_tunnel_send_and_wait_for_l2_confirmation(self, time_travel):
+        """Test tunnel waits for L_DATA.con before sending another L_DATA.req."""
+        self.tunnel.transport.send = Mock()
+        self.tunnel.communication_channel = 1
+
+        test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
+        confirmation = KNXIPFrame.init_from_body(
+            TunnellingRequest(
+                communication_channel_id=1,
+                sequence_counter=23,
+                cemi=CEMIFrame.init_from_telegram(
+                    test_telegram, code=CEMIMessageCode.L_DATA_CON
+                ),
+            )
+        )
+        task = asyncio.create_task(self.tunnel.send_telegram(test_telegram))
+        await time_travel(0)
+        assert not task.done()
+        assert self.tunnel.transport.send.call_count == 1
+        self.tunnel.transport.handle_knxipframe(confirmation, HPAI())
+        await time_travel(0)
+        assert task.done()
+        # one call for the outgoing request. No Acks for TCP.
+        assert self.tunnel.transport.send.call_count == 1
+        await task
+
+    async def test_tunnel_heartbeat(self, time_travel):
+        """Test tunnel sends heartbeat frame."""
+        local_addr = ("192.168.1.1", 12345)
+        remote_hpai = HPAI(
+            ip_addr="192.168.1.2", port=3671, protocol=HostProtocol.IPV4_TCP
+        )
+        self.tunnel.transport.connect = AsyncMock()
+        self.tunnel.transport.getsockname = Mock(return_value=(local_addr))
+        self.tunnel.transport.send = Mock()
+        self.tunnel.transport.stop = Mock()
+        self.tunnel._tunnel_lost = Mock()
+
+        # Connect
+        connection_task = asyncio.create_task(self.tunnel.connect())
+        await time_travel(0)
+        self.tunnel.transport.connect.assert_called_once()
+
+        connect_response_frame = KNXIPFrame.init_from_body(
+            ConnectResponse(
+                communication_channel=23,
+                data_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+                identifier=7,
+            )
+        )
+        self.tunnel.transport.handle_knxipframe(connect_response_frame, remote_hpai)
+        await connection_task
+        self.tunnel.transport.send.reset_mock()
+
+        # Send heartbeat
+        heartbeat_request = KNXIPFrame.init_from_body(
+            ConnectionStateRequest(
+                communication_channel_id=23,
+                control_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+            )
+        )
+        heartbeat_response = KNXIPFrame.init_from_body(
+            ConnectionStateResponse(
+                communication_channel_id=23,
+                status_code=ErrorCode.E_NO_ERROR,
+            )
+        )
+        await time_travel(HEARTBEAT_RATE)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(heartbeat_response, remote_hpai)
+        # test no retry-heartbeat was sent
+        await time_travel(CONNECTIONSTATE_REQUEST_TIMEOUT)
+        self.tunnel.transport.send.assert_not_called()
+        # next regular heartbeat
+        await time_travel(HEARTBEAT_RATE - CONNECTIONSTATE_REQUEST_TIMEOUT)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel._tunnel_lost.assert_not_called()
+
+    async def test_tunnel_heartbeat_no_answer(self, time_travel):
+        """Test tunnel sends heartbeat frame."""
+        local_addr = ("192.168.1.1", 12345)
+        remote_hpai = HPAI(
+            ip_addr="192.168.1.2", port=3671, protocol=HostProtocol.IPV4_TCP
+        )
+        self.tunnel.transport.connect = AsyncMock()
+        self.tunnel.transport.getsockname = Mock(return_value=(local_addr))
+        self.tunnel.transport.send = Mock()
+        self.tunnel.transport.stop = Mock()
+        self.tunnel._tunnel_lost = Mock()
+
+        # Connect
+        connection_task = asyncio.create_task(self.tunnel.connect())
+        await time_travel(0)
+        self.tunnel.transport.connect.assert_called_once()
+
+        connect_response_frame = KNXIPFrame.init_from_body(
+            ConnectResponse(
+                communication_channel=23,
+                data_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+                identifier=7,
+            )
+        )
+        self.tunnel.transport.handle_knxipframe(connect_response_frame, remote_hpai)
+        await connection_task
+        self.tunnel.transport.send.reset_mock()
+
+        # Send heartbeat
+        heartbeat_request = KNXIPFrame.init_from_body(
+            ConnectionStateRequest(
+                communication_channel_id=23,
+                control_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+            )
+        )
+        await time_travel(HEARTBEAT_RATE)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        # no answer - repeat 3 times
+        await time_travel(CONNECTIONSTATE_REQUEST_TIMEOUT)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        await time_travel(CONNECTIONSTATE_REQUEST_TIMEOUT)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        await time_travel(CONNECTIONSTATE_REQUEST_TIMEOUT)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        await time_travel(CONNECTIONSTATE_REQUEST_TIMEOUT)
+        # no answer - tunnel lost
+        self.tunnel._tunnel_lost.assert_called_once()
+
+    async def test_tunnel_heartbeat_error(self, time_travel):
+        """Test tunnel sends heartbeat frame."""
+        local_addr = ("192.168.1.1", 12345)
+        remote_hpai = HPAI(
+            ip_addr="192.168.1.2", port=3671, protocol=HostProtocol.IPV4_TCP
+        )
+        self.tunnel.transport.connect = AsyncMock()
+        self.tunnel.transport.getsockname = Mock(return_value=(local_addr))
+        self.tunnel.transport.send = Mock()
+        self.tunnel.transport.stop = Mock()
+        self.tunnel._tunnel_lost = Mock()
+
+        # Connect
+        connection_task = asyncio.create_task(self.tunnel.connect())
+        await time_travel(0)
+        self.tunnel.transport.connect.assert_called_once()
+
+        connect_response_frame = KNXIPFrame.init_from_body(
+            ConnectResponse(
+                communication_channel=23,
+                data_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+                identifier=7,
+            )
+        )
+        self.tunnel.transport.handle_knxipframe(connect_response_frame, remote_hpai)
+        await connection_task
+        self.tunnel.transport.send.reset_mock()
+
+        # Send heartbeat
+        heartbeat_request = KNXIPFrame.init_from_body(
+            ConnectionStateRequest(
+                communication_channel_id=23,
+                control_endpoint=HPAI(protocol=HostProtocol.IPV4_TCP),
+            )
+        )
+        heartbeat_error_response = KNXIPFrame.init_from_body(
+            ConnectionStateResponse(
+                communication_channel_id=23,
+                status_code=ErrorCode.E_CONNECTION_ID,
+            )
+        )
+        await time_travel(HEARTBEAT_RATE)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(heartbeat_error_response, remote_hpai)
+        # repeat 3 times
+        await time_travel(0)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(heartbeat_error_response, remote_hpai)
+        await time_travel(0)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(heartbeat_error_response, remote_hpai)
+        await time_travel(0)
+        self.tunnel.transport.send.assert_called_once_with(heartbeat_request)
+        self.tunnel.transport.send.reset_mock()
+        self.tunnel.transport.handle_knxipframe(heartbeat_error_response, remote_hpai)
+        await time_travel(0)
+        # third retry had an error - tunnel lost
+        self.tunnel._tunnel_lost.assert_called_once()